25/12/2020
reddit: seperated api from reddit, for other cogs just in case
